+++ generators/app/index.js
@@ -7,6 +7,7 @@ module.exports = class AppGenerator extends EngageGenerator {
     this.composeWith(require.resolve("../client"), {
       ...this.options,
       generateInto: path.join(this.options.generateInto, "client"),
+      "output-dir": "../server/public/dist",
     });
     this.composeWith(require.resolve("../server"), {
       ...this.options,