+++ generators/client/templates/webpack.config.js
@@ -1,30 +1,95 @@
 const path = require("path");
 const webpack = require("webpack");
 
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+
+const isDevelopment = (process.env.NODE_ENV || "development") === "development";
+
 module.exports = {
   entry: ["./src/index.js"],
-  mode: "development",
+  mode: isDevelopment ? "development" : "production",
+  plugins: [
+    new webpack.HotModuleReplacementPlugin(),
+    new MiniCssExtractPlugin({
+      filename: isDevelopment ? "[name].css" : "[name].[hash].css",
+      chunkFilename: isDevelopment ? "[id].css" : "[id].[hash].css",
+    }),
+  ],
   module: {
     rules: [
       {
         test: /\.(js|jsx)$/,
         exclude: /(node_modules|bower_components)/,
         loader: "babel-loader",
-        options: { presets: ["@babel/env"] }
-      }
-    ]
+        options: { presets: ["@babel/env"] },
+      },
+      {
+        test: /\.(png|jpe?g|gif)$/i,
+        use: [
+          {
+            loader: "file-loader",
+          },
+        ],
+      },
+      {
+        test: /\.module\.s(a|c)ss$/,
+        loader: [
+          isDevelopment ? "style-loader" : MiniCssExtractPlugin.loader,
+          {
+            loader: "css-loader",
+            options: {
+              modules: true,
+              sourceMap: isDevelopment,
+              esModule: true,
+              hmr: isDevelopment,
+            },
+          },
+          {
+            loader: "sass-loader",
+            options: {
+              sourceMap: isDevelopment,
+            },
+          },
+        ],
+      },
+      {
+        test: /\.s(a|c)ss$/,
+        exclude: /\.module.(s(a|c)ss)$/,
+        loader: [
+          isDevelopment ? "style-loader" : MiniCssExtractPlugin.loader,
+          "css-loader",
+          {
+            loader: "sass-loader",
+            options: {
+              sourceMap: isDevelopment,
+            },
+          },
+        ],
+      },
+    ],
   },
-  resolve: { extensions: ["*", ".js", ".jsx"] },
+  resolve: {
+    alias: {
+      "@Components": path.resolve(__dirname, "src/components/"),
+      "@Providers": path.resolve(__dirname, "src/providers/"),
+    },
+    extensions: ["*", ".js", ".jsx", ".scss"],
+  },
   output: {
-    path: path.resolve(__dirname, "dist/"),
+    path: path.resolve(__dirname, "<%= options["output-dir"] %>"),
     publicPath: "/dist/",
-    filename: "bundle.js"
+    filename: "bundle.js",
   },
   devServer: {
     contentBase: path.join(__dirname, "public/"),
     port: 3000,
     publicPath: "http://localhost:3000/dist/",
-    hotOnly: true
+    hotOnly: true,
+    proxy: [
+      {
+        context: ["/auth", "/api"],
+        target: "http://localhost:4000",
+      },
+    ],
   },
-  plugins: [new webpack.HotModuleReplacementPlugin()]
 };